apps/ollyscale/tests/test_alerts.py
# BSD 3-Clause License
#
# Copyright (c) 2025, Infrastructure Architects, LLC
# All rights reserved.
#
# Redistribution and use in source and binary forms, with or without
# modification, are permitted provided that the following conditions are met:
#
# 1. Redistributions of source code must retain the above copyright notice,
#    this list of conditions and the following disclaimer.
#
# 2. Redistributions in binary form must reproduce the above copyright notice,
#    this list of conditions and the following disclaimer in the documentation
#    and/or other materials provided with the distribution.
#
# 3. Neither the name of the copyright holder nor the names of its
#    contributors may be used to endorse or promote products derived from this
#    software without specific prior written permission.
#
# THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
# AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
# IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
# DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE
# FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
# DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
# SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
# CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
# OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
# OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

"""
Tests for alert storage and management.

These tests verify:
- Alert creation and storage
- Alert state transitions
- Alert acknowledgment
- Alert queries
"""

import pytest


class TestAlertStorage:
    """Tests for alert storage operations."""

    @pytest.mark.asyncio
    async def test_create_alert(self, storage):
        """Test creating a new alert."""
        pytest.skip("SQLite storage not yet implemented")

    @pytest.mark.asyncio
    async def test_get_alert_by_id(self, storage):
        """Test retrieving alert by ID."""
        pytest.skip("SQLite storage not yet implemented")

    @pytest.mark.asyncio
    async def test_list_alerts(self, storage):
        """Test listing all alerts."""
        pytest.skip("SQLite storage not yet implemented")

    @pytest.mark.asyncio
    async def test_list_active_alerts(self, storage):
        """Test listing only active alerts."""
        pytest.skip("SQLite storage not yet implemented")


class TestAlertStateTransitions:
    """Tests for alert state management."""

    @pytest.mark.asyncio
    async def test_alert_firing_to_resolved(self, storage):
        """Test transitioning alert from firing to resolved."""
        pytest.skip("SQLite storage not yet implemented")

    @pytest.mark.asyncio
    async def test_alert_acknowledge(self, storage):
        """Test acknowledging an alert."""
        pytest.skip("SQLite storage not yet implemented")

    @pytest.mark.asyncio
    async def test_alert_silence(self, storage):
        """Test silencing an alert."""
        pytest.skip("SQLite storage not yet implemented")

    @pytest.mark.asyncio
    async def test_alert_refire(self, storage):
        """Test alert firing again after resolution."""
        pytest.skip("SQLite storage not yet implemented")


class TestAlertQueries:
    """Tests for alert querying."""

    @pytest.mark.asyncio
    async def test_filter_by_severity(self, storage):
        """Test filtering alerts by severity."""
        pytest.skip("SQLite storage not yet implemented")

    @pytest.mark.asyncio
    async def test_filter_by_service(self, storage):
        """Test filtering alerts by service."""
        pytest.skip("SQLite storage not yet implemented")

    @pytest.mark.asyncio
    async def test_filter_by_time_range(self, storage):
        """Test filtering alerts by time range."""
        pytest.skip("SQLite storage not yet implemented")

    @pytest.mark.asyncio
    async def test_alert_history(self, storage):
        """Test retrieving alert history."""
        pytest.skip("SQLite storage not yet implemented")
